[TASK] move search and demand to DTO
