+++ Classes/Domain/Repository/AbstractDemandRepository.php
@@ -2,7 +2,7 @@
 
 namespace Pixelant\PxaDealers\Domain\Repository;
 
-use Pixelant\PxaDealers\Domain\Model\Demand;
+use Pixelant\PxaDealers\Domain\Model\DTO\Demand;
 use TYPO3\CMS\Extbase\Persistence\QueryInterface;
 use TYPO3\CMS\Extbase\Persistence\Repository;
