+++ Classes/Domain/Repository/CategoryRepository.php
@@ -26,7 +26,7 @@
 
 namespace Pixelant\PxaDealers\Domain\Repository;
 
-use Pixelant\PxaDealers\Domain\Model\Demand;
+use Pixelant\PxaDealers\Domain\Model\DTO\Demand;
 use TYPO3\CMS\Extbase\Persistence\QueryInterface;
 
 /**
