+++ Classes/Domain/Repository/DemandRepositoryInterface.php
@@ -2,7 +2,7 @@
 
 namespace Pixelant\PxaDealers\Domain\Repository;
 
-use Pixelant\PxaDealers\Domain\Model\Demand;
+use Pixelant\PxaDealers\Domain\Model\DTO\Demand;
 
 /**
  * Interface DemandRepositoryInterface
